Extracted stats.py from misc.py.
